ENH: Improvements from review
