+++ contrib/api-apache.conf
@@ -2,14 +2,12 @@
 # SPDX-License-Identifier: CC0-1.0
 
 
-<VirtualHost *:80>
-	# To use not default location, allign the ROOT_PATH in the service configuration
-	<Location /intelmq/>
-		ProxyPass unix:/usr/lib/python3/dist-packages/intelmq_api/intelmq_api.sock|http://127.0.0.1/
-		ProxyPassReverse unix:/usr/lib/python3/dist-packages/intelmq_api/intelmq_api.sock|http://127.0.0.1/
-	</Location>
+# To use not default location, allign the ROOT_PATH in the service configuration
+<Location /intelmq/>
+	ProxyPass unix:/usr/lib/python3/dist-packages/intelmq_api/intelmq_api.sock|http://127.0.0.1/
+	ProxyPassReverse unix:/usr/lib/python3/dist-packages/intelmq_api/intelmq_api.sock|http://127.0.0.1/
+</Location>
 
-   <Directory /usr/lib/python3/dist-packages/intelmq_api/>
-      Require all granted
-   </Directory>
-</VirtualHost>
+<Directory /usr/lib/python3/dist-packages/intelmq_api/>
+	Require all granted
+</Directory>
