+++ intelmq_api/models.py
@@ -1,0 +1,17 @@
+"""Models used in API
+
+SPDX-FileCopyrightText: 2023 CERT.at GmbH <https://cert.at/>
+SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+
+from pydantic import BaseModel
+
+
+class LoginForm(BaseModel):
+    username: str
+    password: str
+
+
+class TokenResponse(BaseModel):
+    login_token: str
+    username: str
