+++ intelmq_api/version.py
@@ -3,5 +3,5 @@
 SPDX-FileCopyrightText: 2020-2022 Birger Schacht, Sebastian Wagner
 SPDX-License-Identifier: AGPL-3.0-or-later
 """
-__version_info__ = (3, 1, 0, 'rc2')
+__version_info__ = (3, 1, 0)
 __version__ = '.'.join(map(str, __version_info__))
